New Features and patches.
Added an NSFW filter to the AI Images browser, added a more efficient GetPrompts.js file, waiting to be utilized, added a new GetComponents module which allows you to get the prompt and stuff of an AI image, moved the LogMatch function for ESports_Projects to a web module, since I'm moving that to the web, added authentication stuff for the new logging system, added new methods for getting the players off of the ESports_Projects, and updated the GetPlayers method to be more readable. I also added a new GetMaxMatches module which gets the person with the most number of matches in ESports_Projects.

Patched:
Image scaling in random garlic bread/mashed potatoes browser

Note:
The CheckPasscode module isn't present in this patch because publishing it would be unsecure... at least for the moment.